feat(classpath): Build and include webapp classpath
Add an assembly to build a webapp level classpath.
This classpath is the app module classpath without the bonita connector
and actor filter included (and their transitive dependencies).

Once generated in the target/classpath folder of the app module, the
jars of this folder are merged into the bonita/WEB-INF/lib folder of the
webapp.